requirements.txt
discord.py
python-dotenv
requests
sqlalchemy
eth-ape